servicedb/requirements.txt
psycopg2-binary
pika
fastapi
pydantic
uvicorn
